Move tests to separate packages so that I can e.g. run all offline tests
