+++ androidTest/com/b44t/messenger/uitests/online/OnboardingTest.java
@@ -1,4 +1,4 @@
-package com.b44t.messenger;
+package com.b44t.messenger.uitests.online;
 
 
 import android.content.Context;
@@ -8,6 +8,8 @@ import android.text.TextUtils;
 import androidx.test.ext.junit.rules.ActivityScenarioRule;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.filters.LargeTest;
+
+import com.b44t.messenger.TestUtils;
 
 import org.junit.After;
 import org.junit.Rule;
